refactor: extract execute_replace_op and execute_search_op from tx/execute.rs (#906)
Move the two largest functions out of tx/execute.rs into dedicated
submodules:
- tx/replace_op.rs (191 lines): full replace operation with glob
  walking, fallback matching, and did-you-mean hints
- tx/search_op.rs (209 lines): search within transactions with
  binary detection, context lines, and assert_count

tx/execute.rs shrinks from 1,363 to 975 lines (under the 1,000-line
target in the issue acceptance criteria). The functions keep their
pub(crate) visibility via re-exports from execute.rs so all existing
call sites work unchanged.
