+++ src/tx/execute.rs
@@ -1,4 +1,4 @@
-use super::output::{TxExecResult, TxLintResult, TxReadResult, TxSearchMatch, TxSearchResult};
+use super::output::{TxExecResult, TxLintResult, TxReadResult, TxSearchResult};
 use super::validate::op_label;
 use crate::cli::global::{EolMode, GlobalFlags};
 use crate::ops::doc::{
@@ -10,14 +10,10 @@ use crate::ops::md::{
     replace_section_in, table_append_for_tx, upsert_bullet_in,
 };
 use crate::ops::patch::{ApplyHunksOptions, ApplyHunksStatus, apply_patch_with_loader};
-use crate::ops::replace::{
-    compile_replace_regex, replace_content, replace_whole_lines, replacement_text,
-};
+
 use crate::plan::{Operation, Plan};
 use crate::write::{WritePolicy, apply_policy};
-use globset::Glob;
-use ignore::WalkBuilder;
-use regex::RegexBuilder;
+
 use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
@@ -229,189 +225,8 @@ pub(crate) struct TxState<'a> {
     pub(crate) structured: bool,
 }
 
-/// Execute a replace operation within a transaction.
-pub(crate) fn execute_replace_op(op: &Operation, tx: &mut TxState<'_>) -> anyhow::Result<usize> {
-    let Operation::Replace {
-        glob,
-        path,
-        mode,
-        from,
-        to,
-        nth,
-        insert_before,
-        insert_after,
-        case_insensitive,
-        multiline,
-        whole_line,
-        range,
-        before_context,
-        after_context,
-        ..
-    } = op
-    else {
-        unreachable!()
-    };
-    let regex_mode = mode.as_deref() == Some("regex");
-    let word_boundary = matches!(
-        op,
-        Operation::Replace {
-            word_boundary: true,
-            ..
-        }
-    );
-    let use_regex = regex_mode || *case_insensitive || word_boundary;
-    let replacement = replacement_text(from, to, insert_before, insert_after, use_regex);
-    let compiled_re = compile_replace_regex(
-        from,
-        regex_mode,
-        *case_insensitive,
-        *multiline,
-        word_boundary,
-    )?;
-    let parsed_range = range
-        .as_deref()
-        .map(crate::ops::read::parse_line_range)
-        .transpose()?;
-
-    if let Some(p) = path {
-        let file_path = tx.cwd.join(p);
-        let content = read_file_content(tx.pending, tx.existed_before, &file_path)?;
-        let (replaced, match_count) = if *whole_line {
-            replace_whole_lines(
-                content,
-                from,
-                &replacement,
-                compiled_re.as_ref(),
-                *nth,
-                parsed_range,
-            )
-        } else {
-            replace_content(content, from, &replacement, compiled_re.as_ref(), *nth)
-        };
-        if match_count > 0 {
-            let owned = replaced.into_owned();
-            update_file_content(tx.pending, tx.deletions, &file_path, owned);
-            Ok(match_count)
-        } else if !regex_mode && (before_context.is_some() || after_context.is_some()) {
-            // Tier 3: Use context-based fallback when exact match fails.
-            match crate::fallback::resolve_with_fallback(
-                content,
-                from,
-                before_context.as_deref(),
-                after_context.as_deref(),
-            ) {
-                Ok(anchor) => {
-                    let to_text = to.as_deref().unwrap_or("");
-                    let new_content = format!(
-                        "{}{}{}",
-                        &content[..anchor.start_offset],
-                        to_text,
-                        &content[anchor.start_offset + anchor.matched_text.len()..]
-                    );
-                    update_file_content(tx.pending, tx.deletions, &file_path, new_content);
-                    tx.replace_hint = Some(format!(
-                        "fallback matched via {:?} strategy in {}",
-                        anchor.strategy, p,
-                    ));
-                    Ok(1)
-                }
-                Err(edit_error) => {
-                    tx.replace_hint = Some(edit_error.message.clone());
-                    Ok(0)
-                }
-            }
-        } else {
-            if !regex_mode {
-                // Tier 1: Provide "did you mean?" hints for literal no-match.
-                let similar = crate::fallback::find_similar_targets(content, from, 3);
-                if !similar.is_empty() {
-                    tx.replace_hint = Some(format!(
-                        "no matches for '{}' in {} (did you mean: {}?)",
-                        crate::fallback::truncate_str(from, 60),
-                        p,
-                        similar.join(", ")
-                    ));
-                }
-            }
-            Ok(0)
-        }
-    } else if let Some(pattern) = glob {
-        let matcher = Glob::new(pattern)?.compile_matcher();
-        let matches_pattern = |path: &Path| {
-            matcher.is_match(path)
-                || path.file_name().is_some_and(|name| matcher.is_match(name))
-                || path.strip_prefix(tx.cwd).ok().is_some_and(|relative| {
-                    !relative.as_os_str().is_empty()
-                        && (matcher.is_match(relative)
-                            || relative
-                                .file_name()
-                                .is_some_and(|name| matcher.is_match(name)))
-                })
-        };
-        let mut total_matches = 0usize;
-        let mut candidate_paths = Vec::new();
-        let mut seen_paths = HashSet::new();
-
-        let walker = WalkBuilder::new(tx.cwd).build();
-        for entry in walker {
-            let entry = match entry {
-                Ok(e) => e,
-                Err(_) => continue,
-            };
-            if !entry.file_type().is_some_and(|ft| ft.is_file()) {
-                continue;
-            }
-            let file_path = entry.path().to_path_buf();
-            if matches_pattern(&file_path) && seen_paths.insert(file_path.clone()) {
-                candidate_paths.push(file_path);
-            }
-        }
-
-        for pending_path in tx.pending.keys() {
-            if !pending_path.starts_with(tx.cwd)
-                || pending_path.exists()
-                || tx.deletions.contains(pending_path)
-                || !matches_pattern(pending_path)
-            {
-                continue;
-            }
-            if seen_paths.insert(pending_path.clone()) {
-                candidate_paths.push(pending_path.clone());
-            }
-        }
-
-        for file_path in candidate_paths {
-            let content = match read_file_content(tx.pending, tx.existed_before, &file_path) {
-                Ok(c) => c.to_owned(),
-                Err(e) => {
-                    if !tx.structured && !tx.quiet {
-                        eprintln!("tx: replace: skipping {}: {e}", file_path.display());
-                    }
-                    continue;
-                }
-            };
-            let (replaced, match_count) = if *whole_line {
-                replace_whole_lines(
-                    &content,
-                    from,
-                    &replacement,
-                    compiled_re.as_ref(),
-                    *nth,
-                    parsed_range,
-                )
-            } else {
-                replace_content(&content, from, &replacement, compiled_re.as_ref(), *nth)
-            };
-            total_matches += match_count;
-            if match_count > 0 {
-                update_file_content(tx.pending, tx.deletions, &file_path, replaced.into_owned());
-            }
-        }
-        Ok(total_matches)
-    } else {
-        anyhow::bail!("replace operation requires either 'path' or 'glob'");
-    }
-}
+// execute_replace_op extracted to tx/replace_op.rs (#906).
+pub(crate) use super::replace_op::execute_replace_op;
 
 /// Execute a read operation within a transaction.
 pub(crate) fn execute_read_op(
@@ -455,211 +270,8 @@ pub(crate) fn execute_read_op(
     Ok(())
 }
 
-/// Execute a search operation within a transaction.
-///
-/// If `path` is a directory, walks it (respecting `.gitignore`) and searches
-/// each non-binary file, mirroring the standalone `search` command behavior.
-pub(crate) fn execute_search_op(op: &Operation, tx: &mut TxState<'_>) -> anyhow::Result<()> {
-    let Operation::Search {
-        path,
-        pattern,
-        regex,
-        case_insensitive,
-        multiline,
-        invert_match,
-        context,
-        before_context,
-        after_context,
-        assert_count,
-        literal,
-        globs,
-        max_results,
-        exclude_patterns,
-        custom_ignore_filenames,
-    } = op
-    else {
-        unreachable!()
-    };
-
-    if *invert_match && *multiline {
-        anyhow::bail!("invert_match and multiline cannot be combined");
-    }
-    if *literal && *regex {
-        anyhow::bail!("search: literal and regex cannot be combined");
-    }
-
-    // literal means treat pattern as literal text (escape it); !regex also escapes for backward compat in plans.
-    let pat = if *literal || !*regex {
-        regex::escape(pattern)
-    } else {
-        pattern.clone()
-    };
-    let re = {
-        let mut builder = RegexBuilder::new(&pat);
-        builder.case_insensitive(*case_insensitive);
-        builder.dot_matches_new_line(*multiline);
-        builder.build()?
-    };
-
-    let ctx_before = before_context.or(*context).unwrap_or(0);
-    let ctx_after = after_context.or(*context).unwrap_or(0);
-
-    let resolved = tx.cwd.join(path);
-    // Use the shared advanced ignore collection when available for parity with
-    // api::search_directory / CLI (respects .gitignore + custom_ignore_filenames + exclude_patterns).
-    // Then apply include globs.
-    #[cfg(any(feature = "cli", feature = "files"))]
-    let candidate_paths: Vec<PathBuf> = if resolved.is_dir() {
-        crate::files::collect_file_paths_with_ignores(
-            &resolved,
-            custom_ignore_filenames,
-            exclude_patterns,
-            false,
-        )?
-    } else {
-        vec![resolved.clone()]
-    };
-    #[cfg(not(any(feature = "cli", feature = "files")))]
-    let candidate_paths: Vec<PathBuf> = if resolved.is_dir() {
-        let mut paths = Vec::new();
-        for entry in WalkBuilder::new(&resolved).build() {
-            let entry = entry?;
-            if entry.file_type().is_some_and(|ft| ft.is_file()) {
-                paths.push(entry.into_path());
-            }
-        }
-        paths.sort();
-        paths
-    } else {
-        vec![resolved.clone()]
-    };
-
-    let glob_matcher = crate::files::build_glob_matcher(globs)?;
-    // glob_roots used for relative glob matching; use the search root (dir or parent of file)
-    let glob_root = if resolved.is_dir() {
-        resolved.clone()
-    } else {
-        resolved.parent().unwrap_or(&resolved).to_path_buf()
-    };
-    let glob_roots = vec![glob_root];
-
-    let file_paths: Vec<PathBuf> = if let Some(m) = &glob_matcher {
-        candidate_paths
-            .into_iter()
-            .filter(|p| crate::files::matches_glob_with_roots(p, Some(m), &glob_roots))
-            .collect()
-    } else {
-        candidate_paths
-    };
-
-    let mut all_matches = Vec::new();
-
-    for file_path in &file_paths {
-        // For directory walks, read the file once and check for binary content
-        // in the same buffer (avoids double-reading text files: 8 KiB probe +
-        // full re-read). For single-file paths, skip the binary check.
-        if file_paths.len() > 1 {
-            if !read_and_probe(tx.pending, tx.existed_before, file_path)? {
-                continue; // binary file, skip
-            }
-        } else {
-            read_file_content(tx.pending, tx.existed_before, file_path)?;
-        }
-        let content = &tx.pending[file_path].1;
-        let lines: Vec<&str> = content.lines().collect();
-
-        let is_multi_file = file_paths.len() > 1;
-        if *multiline {
-            // Multiline mode: search the full content so patterns can span lines.
-            for m in re.find_iter(content) {
-                let line_idx = content[..m.start()].matches('\n').count();
-                let start = line_idx.saturating_sub(ctx_before);
-                let end = (line_idx + 1 + ctx_after).min(lines.len());
-                let matched_text = m.as_str().to_string();
-                let text = if is_multi_file {
-                    let display_path = file_path
-                        .strip_prefix(tx.cwd)
-                        .unwrap_or(file_path)
-                        .to_string_lossy();
-                    format!("{display_path}:{matched_text}")
-                } else {
-                    matched_text
-                };
-                let col = m.start() - content[..m.start()].rfind('\n').map_or(0, |p| p + 1);
-                all_matches.push(TxSearchMatch {
-                    line: line_idx + 1,
-                    column: col + 1,
-                    text,
-                    context_before: lines[start..line_idx]
-                        .iter()
-                        .map(|s| s.to_string())
-                        .collect(),
-                    context_after: lines[line_idx + 1..end]
-                        .iter()
-                        .map(|s| s.to_string())
-                        .collect(),
-                });
-            }
-        } else {
-            for (i, line) in lines.iter().enumerate() {
-                let found = re.find(line);
-                let is_match = if *invert_match {
-                    found.is_none()
-                } else {
-                    found.is_some()
-                };
-                if is_match {
-                    let start = i.saturating_sub(ctx_before);
-                    let end = (i + 1 + ctx_after).min(lines.len());
-                    let text = if is_multi_file {
-                        let display_path = file_path
-                            .strip_prefix(tx.cwd)
-                            .unwrap_or(file_path)
-                            .to_string_lossy();
-                        format!("{display_path}:{}", line)
-                    } else {
-                        line.to_string()
-                    };
-                    let column = found.map_or(1, |m| m.start() + 1);
-                    all_matches.push(TxSearchMatch {
-                        line: i + 1,
-                        column,
-                        text,
-                        context_before: lines[start..i].iter().map(|s| s.to_string()).collect(),
-                        context_after: lines[i + 1..end].iter().map(|s| s.to_string()).collect(),
-                    });
-                }
-            }
-        }
-    }
-
-    // Cap after collection (and the inner loop has already sorted? no, tx does not sort here; append order is walk order).
-    // For determinism in report we could sort, but to minimize diff keep original order and just truncate.
-    // (CLI sorts for output; tx report order is from walk.)
-    if *max_results > 0 {
-        all_matches.truncate(*max_results);
-    }
-
-    if let Some(expected) = assert_count
-        && all_matches.len() != *expected
-    {
-        anyhow::bail!(
-            "search assert_count: expected {} matches for '{}' in {}, found {}",
-            expected,
-            pattern,
-            path,
-            all_matches.len()
-        );
-    }
-
-    tx.tx_searches.push(TxSearchResult {
-        path: path.clone(),
-        pattern: pattern.clone(),
-        match_count: all_matches.len(),
-        matches: all_matches,
-    });
-    Ok(())
-}
+// execute_search_op extracted to tx/search_op.rs (#906).
+pub(crate) use super::search_op::execute_search_op;
 
 /// Returns true if the operation works on raw text and needs any cached
 /// doc values flushed (serialized) back to the pending text map first.
